Overwrite error
When testing on Linux, the *.tab output file only contained the initial line of text with molecule names. Temporary solution was to uncomment `!CALL sysclose(fu_plt, pltnam, error)` at the end of ops_plot_uitv.

Actual cause of the problem was found to be overwriting under specific conditions on Linux.